feat: convert visit type to dropdown in edit visit
